+++ gradient_field.hpp
@@ -17,14 +17,14 @@ class gradient_field {
 
         void print_field();
         void warp_field();
-        
-        float get_height(int pos_x, int pos_y);
+
+        float get_height(float pos_x, float pos_y);
         float perlin_noise(float point_x, float point_y);
         float smoothstep(float vec_1, float vec_2, float weight);
         float distance_scalar_product(int ix, int iy, float x, float y);
 
-        float map_to_field_x(int point);
-        float map_to_field_y(int point);
+        float map_to_field_x(float point);
+        float map_to_field_y(float point);
         // float map_y(int point);
 
         int positive_mod(int dividend, int divisor);
